# Regionprops table performance refactor