@pep8speaks 2021-09-23
Hello @grlee77! Thanks for opening this PR. We checked the lines you've touched for [PEP 8](https://www.python.org/dev/peps/pep-0008) issues, and found:

* In the file [`skimage/measure/_regionprops.py`](https://github.com/scikit-image/scikit-image/blob/890b1538ad7e2c75cd8041b47fe6893cd9f33f54/skimage/measure/_regionprops.py):

> [Line 270:80](https://github.com/scikit-image/scikit-image/blob/890b1538ad7e2c75cd8041b47fe6893cd9f33f54/skimage/measure/_regionprops.py#L270): [E501](https://duckduckgo.com/?q=pep8%20E501) line too long (83 > 79 characters)